+++ test/e2e/steps/taskCreation.steps.js
@@ -1,56 +1,49 @@
-const { When } = require('cucumber');
-const LoginPage = require('../pages/LoginPage');
-const DashboardPage = require('../pages/DashboardPage');
-const TaskPage = require('../pages/TaskPage');
 const BoardPage = require('../pages/BoardPage');
-const Utils = require('../common/Utils');
+const Context = require('../../data/Context');
+const DashboardPage = require('../pages/DashboardPage');
+const LoginPage = require('../pages/LoginPage');
 const SystemInteractions = require('../constants/SystemInteractions');
+const TaskAsserts = require('../asserts/Task.assert');
+const TaskPage = require('../pages/TaskPage');
+const Utils = require('../common/Utils');
+const { Given, When, Then } = require('cucumber');
 
-When('user fills the fields with the following information:', function (dataTable) {
-  let userData = dataTable.hashes();
-  LoginPage.setEmailInput(userData[0].mail);
-  LoginPage.setPasswordInput(userData[0].password);
-  LoginPage.submitButton.click();
+Given(/^user is in the main page$/, function () {
+  LoginPage.open(Context.page.endpoints.base);
+  browser.deleteCookies();
 });
 
-When('user creates {string} list', function (listName) {
+When(/^user creates "([^"]*)" list$/, function (listName) {
   DashboardPage.addListButton.waitForExist(3000);
   DashboardPage.addListButton.click();
   DashboardPage.newListButton.click();
   DashboardPage.setNewListInput(listName);
   browser.keys(SystemInteractions.ENTER_KEY_PRESS);
 });
-
-When('user creates tasks with the following information:', function (dataTable) {
+  
+When(/^user creates tasks with the following information$/, function (dataTable) {
   let task = dataTable.hashes();
   for(let i = 0; i < task.length; i++) {
     DashboardPage.setTaskName(task[i].taskName);
     DashboardPage.saveTaskButton.click();
   }
 });
-
-When('user clicks on BOARD tab', function () {
+When(/^user clicks on BOARD tab$/, function () {
   DashboardPage.boardTab.waitForExist(3000);
   DashboardPage.boardTab.click();
 });
-
-When('user closes the first task', function () {
+  
+When(/^user closes the first task$/, function () {
   TaskPage.taskTitle.waitForExist(3000);
   TaskPage.taskTitle.moveTo();
   TaskPage.closeTaskIcon.click();
 });
-
-When('user creates {string} board', function (boardName) {
-  DashboardPage.setTaskName('      ');
-  DashboardPage.boardTab.click();
-  DashboardPage.addStatusButton.click();
-  BoardPage.setStatusNameInput(boardName);
-  browser.keys(SystemInteractions.ENTER_KEY_PRESS);
-  DashboardPage.modalText.click();
-  DashboardPage.confirmNewStatus.click();
+  
+Then(/^"([^"]*)" should be in TO DO board and "([^"]*)" in COMPLETE board$/, function (taskName1, taskName2) {
+  TaskAsserts.taskInBoard(taskName1, taskName2);
 });
 
-When('user creates tasks in the following order:', function (dataTable) {
+When(/^user creates tasks in the following order$/, function (dataTable) {
   let tasks = dataTable.hashes();
   for(let i = 0; i < tasks.length; i++) {
     BoardPage.createTaskButton(1).click();
@@ -64,7 +57,7 @@ When('user creates tasks in the following order:', function (dataTable) {
   Utils.tasksFromTable(tasks, 'IN_PROGRESS');
 });
 
-When('user adds description to a given card:', function (dataTable) {
+When(/^user adds description to a given card$/, function (dataTable) {
   let desc = dataTable.hashes();
   desc.forEach(elm => {
     TaskPage.getTaskTitle(elm.CARD).click();
@@ -74,7 +67,7 @@ When('user adds description to a given card:', function (dataTable) {
   });
 });
 
-When('user deletes the following cards:', function (dataTable) {
+When(/^user deletes the following cards$/, function (dataTable) {
   let tasks = dataTable.raw();
   tasks.forEach(elm => {
     TaskPage.getTaskTitle(elm).click();
@@ -82,3 +75,9 @@ When('user deletes the following cards:', function (dataTable) {
     TaskPage.deleteButton.click();
   });
 });
+
+Then(/^the following tasks should be visible$/, function (dataTable) {
+  let tasks = dataTable.raw();
+  tasks.forEach(elm => TaskAsserts.assertTaskExist(elm));
+});
+
